mips: Optimize jit_andi further
On MIPS r2+, when the mask is a high mask (consecutive bits set to 1,
starting at MSB), a middle mask (consecutive bits set to 1 in the
middle of the mask), or an inverted middle mask (consecutive bits set to
1 starting at MSB, consecutive bits set to 1 starting at LSB, bits set
to zero somewhere in the middle), it is possible to simplify the code
generation by using the EXT and INS opcodes and their variants.

This commit adds extr() and insr() helpers, which will use the right
opcodes according to the mask that will be used.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>