Fix configure script's handling of disassembler
The configure script was correctly checking for the necessary
dependencies for building with the disassembler enabled, but
when dependencies were not found, its messages were imprecise
and sometimes incorrect.

The configure script now shows a good message if any of the
following dependencies are not found:

* libiberty
* binutils BFD
* libz
* binutils opcodes

As before, if the '--enable-disassembler' flag is passed, failure to
detect any one of these dependencies is treated as an error.

Also deleted old 'FIXME' comment relating to this.

Signed-off-by: Max Barraclough <maxbarraclough@outlook.com>