Add bswapr_us, bswapr_ui, and bswapr_ul API functions
The difference between bswapr and htonr/ntohr, is that the operation
performed by bswap does not depend on the endianness of the host CPU.

On little-endian however, the new bswapr functions are functionally
equal to the htonr/ntohr ones. The implementations of the latter ones
have then be removed, and will resolve to the corresponding bswapr
function on little-endian, or to movr/extr functions on big-endian.

Optional generic bswapr implementations are provided in lightning.c,
which can be used (or not) by the arch cpu.c files when there's nothing
better to do.

The bswapr implementations are generally following the previous htonr
implementations on little-endian systems, with the exception of the
IA-64 architecture, where the htonr implementation seemed sub-par, and
a better (yet untested) implementation has been added.

This change has been tested on MIPS, x86 and PowerPC.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>