jit_size: Don't round up the size to a multiple of 4 KiB
When using an external code buffer, a program using Lightning will call
jit_get_code() to get the expected size of the code, then allocate some
space in the code buffer, then call jit_set_code() to specify where the
code should be written to.

If the reported size is rounded up to a multiple of 4 KiB, then the
allocator will always try to allocate 4 KiB for blocks of code that
might even be smaller than a hundred bytes. The program can then choose
to realloc() the allocated block to the actual size of the generated
code, to reduce the memory used.

However, this will cause dramatic memory fragmentation; for instance,
if working with a 2 MiB code buffer, in which 512 blocks of 4 KiB are
allocated but later realloc'd to 128 bytes each, the total amount of
allocated memory will be 128 * 512 == 64 KiB, with almost 1.9 MiB free,
yet it will be impossible to allocate any new blocks as there would be
no way to find a contiguous 4 KiB area.

Besides, I really don't understand why it was rounded up to a multiple
of 4 KiB, as this is not a requirement for mmap().

Signed-off-by: Paul Cercueil <paul@crapouillou.net>