ppc: Add optimization to generate endian-swapping load opcodes
When generating endian-swap code with jit_bswapr_us() or
jit_bswapr_ui(), look at the last opcode generated. If it's a LHZX or
LWZX, it can be replaced with a LHBRX or LWBRX opcode, which will
perform the endian swap directly.

If it's a LHZ or LWZ, then we can just move the immediate value into a
temporary register, and emit a LHBRX or LWBRX opcode.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>