Correct wrong code generation when redundant_store removes a movi.
  The pattern was:

    beqi L2 %v0 0
L1:
    jmpi somewhere
L2:
    <instruction list>
    movi %v0 0

  The problem as that the removal of the movi was unsafe, because
<instruction list> could use it as a temporary.
  Now instead of generating a redundant reload of the register value
it modifies the node to tell the register is live, even if it is not
used after that point. So, it could be a pessimization in the rare
case of the movi being in a non callee save register in a dead code
path, but that is another problem, and could be fixed with simple
analysis of the generated code.