Implement hardware casr and casi.
Actually, at least in my https://wiki.debian.org/DebianYeeloong
it does not work in the fallback mode. This should be due to very
old hardware and somehow pthread_mutex_{un,}lock being way too
slow (hardware watch in gdb is not usable...). But the hardware
implementation works.
It did need some extra patches to fallback.c as it uses _AT as a
valid register for jit_get_reg(), that also is not in JIT_R(n) nor
JIT_V(n), so, would need to be saved/restored for casi.