Correct regression in new live block computation logic.
  Basically it did not follow changes to source blocks in some cases
when computing live state. This was due to incorrect assumption that
it would find a cycle when updating backwards the live state, but in
a condition with a jump to a raw address, and the next label having a
single source, if the source live state was changed, it would not
properly forward the live state to a jump target.