Correct another case the live state is not propagated.
  This commit corrects an issue where the live state is not propagated
back and a live register (after several blocks and jumps) is used as a
temporary.
  The live state must be propagated backwards on normal flow, as done
in this change because it might be changed in a block, that will have
the register as dead at entry, then, after several blocks it is used
again. The normal/previous code did not trigger it because it would
fail to follow the flow when there was a jmpi to raw address or jmpr.