Prevent incorrect detection of out of registers
  This change prevents an assertion, but code generation can be made better
for other conditions.
  The problem can be noticed in check/put.tst, that basically touchs all
registers in a block, and then does a jmpi to a location that can be
followed.
  It must reset state, as per the start of the next block before the jmpi,
because the jmpi might attempt to allocate a register that cannot be spilled.
In this condition it might incorrect understand there are no free registers.
  The code should be further modified. Currently it is not broken, but
might cause unnecessary spill/reloads in some conditions, like exercised
in check/put.tst.