Correct code ignoring jit_live and explicit live registers
  Previous change to assume all callee save registers are live at
blocks only reachable with indirect jumps (including functions)
caused the possibility of incorrectly assuming all registers are
live at a jump to a known target, if all registers were touch.
  This condition should happen for all ports in the check/put.tst.
  The correction was to detect the condition, and reset the bitmasks
before a possible need for a register to store the target address on
a jmpi.
  But that caused two side effects, now corrected:
o Just setting the return register as live in the normal path cannot
  be understood, so, now check it explicitly in a function epilog.
o Explicit calls to jit_live were ignored if the block did end in a
  jump to a known target (a label in jit_code). To correct this issue
  a new regset was added to record explicit jit_live calls.

  Note that the special code is only for jmpi. Still need to fully
validate conditional branches on other architectures, as the conditions
were only tested on x86.