Correct inconsistencies when restarting emitting a function.
  Emitting a function body can be restarted if some internal states change.
These are the function->self.aoff and function->regset.
  The aoff field is the alloca offset, that might be implicitly modified
when allocating stack space for a register spill/reload.
  The regset field might be modified when it notices at some point a callee
save register has been used as a temporary, so, the prolog/epilog must be
updated.

  The conditions did not show earlier because when/if a function was
restarted in the tests, the invalid state did not interfere in make check.
  The problem was detected, due to the extra register pressure added in
recent patches to avoid issues with blocks only reachable with indirect
jumps, and in the arm port, where registers are a scarce resource, with
problem noticed with a _vfp_movi_d (that allocates two gpr registers)
causing an exception in check/put.tst due to running out of registers.