mips: Fill delay slots of JALR opcodes in jit_callr
Fill the delay slot of the generated JALR opcode, if it is not already
used to set the $t9 register.

When we know that the last generated opcode is not the target of a jump,
and that it does not write the register used by the JALR opcode, we can
swap it with the JALR opcode, so that it now becomes the delay slot of
the JALR opcode.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>