arm: Optimize procedure calls in the same jit buffer
  The code by default assumes functions in arm mode might call functions
in thumb mode or vice versa. To disable this feature must rebuild
lightning and change jit_exchange_p() to return 0. This will prevent
adding a prolog header to switch to thumb mode. This is only meaningful
if a full thumb mode environment, and only generating (default, unless
thumb instruction set not available) jit with thumb instructions.
  In this change, a shortcut is added to skip the prolog header to
switch from arm to thumb mode. Also, in arm mode it does short branches
if calling a function in the same jit buffer (this was a pessimization
as only jit functions in the same buffer were called in an expensive
mode before this change).