arm: Take advantage of the need_return field for jit functions
  This is done only for jit functions that only call jit_functions. For
now this is a hardcoded safe mode assumption that there might be issues
if a jit function does not create a frame pointer and calls a C function.

  Note that there is also the possible case of passing a jit function
pointer to be called from a C function. It is assumed that in this case
it is safe to have a leaf or optimized function as done in the patch.