mips: Rewrite code to add an out of order instruction in delay slot
  This change adds a complete (in the sense of instructions used by
Lightning) decoder in the jit_get_reg_for_delay_slot() function, that
serves double usage, that are getting a safe temporary register,
and making sure delay slot can be filled with the 'pending' instruction,
otherwise it just flushes code generation.

  There is an order or calls, where the new function pending() cannot
be called before jit_get_reg_for_delay_slot(), as the later might
call flush(), to emit the instruction.

  Code has been refactored to use the helpers to fill the delay slot,
and also get the 'pending' instruction that is more likely to be safe
to be executed out of order. There is no search for already emitted
instructions, only a check shortly before the branch, to avoid yet
more complex code to ensure the instruction order is correct.

  The call to pending(), after the call to jit_get_reg_for_delay_slot()
will return either an instruction that can be executed out of order, or
a nop, that must be added to the delay slot, using the delay*() function.

	* include/lightning/jit_private.h: Add new 'inst' field to
	jit_compiler_t, if __mips__ is defined. This field is a simple
	helper for a pending instruction to be emitted, and that can
	be emitted out of order.
	* lib/jit_fallback.c: Update for changes in internal mips patching
	and jumping macros and function calls.
	* lib/jit_mips-cpu.c: Core of changes to attempt to fill delay
	slots with instructions that can be emitted out of order.
	* lib/jit_mips-fpu.c: Update to use delay slot in branches.
	* lib/jit_mips.c: Update for new delay slot use logic.