s390: build and pass all tests.
  It is not possible to use IC or ICM for jit_dep as is. It must read
from a memory address, so, it would be usable only on multi instruction
optimization.

  The s390 port also shows jit_depi() actually can be optimized, using
IIHF, IIHH, IIHL, IILF, IILH and IILL.  But before this, would be far
more important to optimize load of constants and andi/ori, that can use
less instructions, and are currently badly implemented (works, but can
be shorter).