mips: Implement unaligned memory access.
  Note that Linux kernels should trap and handle unaligned memory access.
  For the moment it does not rely on this behavior by default. To change
it, add the C code:

	jit_cpu.unaligned = 1;

after calling init_jit(). Note that this is only an option if
jit_cpu.version is 5 or lower, as mips release 6 or newer removes the
instructions to use unaligned memory access. This has not yet been fully
optimized for mips 6. The logic should be to construct special instructions
for smaller loads or stores and load/store the unaligned value without
trapping.