arm: Implement movi_w_f and movi_ww_d
  Also add a patch to ensure constants are flushed before JIT_INSTR_MAX
or 128 bytes if JIT_INSTR_MAX is not defined.
  This change still does not fix a temporary issue with incorrect calculation
of buffer size, but once instruction size is redone the problem will go away.
  It will cause an assertion if running in arm mode, disabling vfp and
telling it is arm version 4, when running check/lightning on check/unldst.tst.
(A minor temporary quick hack to have 256 pending bytes fixes it for a local
test, but soon it will no longer be an issue)