Implement the new fnma* and fnms* instructions
  These complete the fused multiply add/sub abstractions.

  There is no work on precise ordering on fallbacks. The interface only
makes available the instructions, and the fallbacks are not guaranteed
to generate the same result for all inputs. This happens because when
actually implemented in hardware, any rounding is only done at the end.
Lightning does not handle floating point exceptions or rounding modes, so,
handling anything other than "default" configuration is up to the user.