hppa: Add but do not use workarounds to pass all tests
  The unaligned load store tests only pass if not using instructions
with an immediate index offset. Even on a build where such instructions
are only used if the final address is guaranteed wordsize aligned it
still fails. For the moment have a runtime modifier, by setting
    jit_cpu.imm_idx = 0
and causing it to always move immediate indexes to a register.
  This is not enabled because the same patterns and encodings pass in
all other tests involving indexed loads and stores.

  The change, also disabled, to catomic.c appears to handle a case where
it crashes acquiring an stdio lock. To avoid the crash do not use stdio.
Because it could also be hidding a bug, it is not enabled. This is not
a runtime change, as it is specific to this single test.
  These should be only debugged further if there is access to actual
hardware.