lightning: Only optimize movr for regular registers
When a jit_movr() or jit_movi() targets a register that is not used by
Lightning, the opcode should be emitted or undefined behaviour may
occur.

This makes it possible to read/write system registers using jit_movr(),
for instance the GBR register on SuperH processors, which is generally
used as the base address of the thread-local storage data.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>