mips: Correct and optimize jit_rshr() and jit_rshr_u()
Rework the branch-less path to shrink the size of jit_rshr() and
jit_rshr_u(), and use one register less, so that the whole code path
that uses branches can be dropped.

The case where O4 == __WORDSIZE in jit_rshr() was also handled
incorrectly, as it would zero the O1 register instead of sign-extending
it.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>