aarch: Implement optimized pre increment load/store
  There are no instructions with a variable increment. The field is used
for other features, like scaling.
  It was noticed load/store of float values use transfer to integer registers
and integer load/stores. This was from the initial code more than 10 years
ago, when documentation was difficult to get. For the moment added a note
that need to update to use simd instructions as these would reduce number
of instructions for normal float load/store as well.