SH4: Rework const flushing
It previously checked the PC vs. a previous const flush, and flushed
again if the delta was >= 450 bytes. This made no sense because the
mov.l opcodes only accept a positive displacement.

Besides, the +450 bytes displacement only makes sense if we plan to use
the mov.w opcode (whose displacement is max. +510 bytes); but it is
generally better to build 16-bit values in code, so mov.w won't be used
to load 16-bit constants. As such, the maximum displacement is now +1020
bytes, so the cutoff is now set to +950 bytes.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>