SH4: Invert FPU registers order
The only point of this is to reduce the chances of FR0 being in use, in
order to generate better code when FMAC is used.

Note that the jit_f() macro was wrong before, as it would list the
registers in reverse order. This caused the "self" check to fail.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>